feat: move the search button
